pyproject.toml
[project]
name = "try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "quixstreams>=3.21.0",
    "websocket-client>=1.8.0",
]

[tool.uv.workspace]
members = [
    "serivces/trades",
]
